@@ -3,29 +3,24 @@
   "version": "1.3.4",
   "description": "The Quill rich-text editor as a React component.",
   "author": "zenoamaro <zenoamaro@gmail.com>",
-  "homepage": "https://github.com/zenoamaro/react-quill",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/zenoamaro/react-quill/issues",
-    "email": "zenoamaro@gmail.com"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/zenoamaro/react-quill.git"
-  },
+  "homepage": "https://github.com/zenoamaro/react-quill",
+  "bugs": "https://github.com/zenoamaro/react-quill/issues",
+  "repository": "https://github.com/zenoamaro/react-quill.git",
   "main": "lib/index.js",
-  "browser": "lib/index.js",
-  "types": "types.d.ts",
+  "types": "lib/index.d.ts",
   "scripts": {
-    "test": "make test",
-    "build": "make build",
-    "prepack": "make build",
-    "postpack": "make clean",
-    "prepublish": "make build",
-    "postpublish": "make clean"
-  },
-  "engines": {
-    "node": ">= 0.8.x"
+    "build": "npm run build:lib && npm run build:dist && npm run build:css",
+    "build:lib": "tsc",
+    "build:dist": "webpack",
+    "build:css": "cpx 'node_modules/quill/dist/quill.*.css' dist",
+    "watch": "tsc --watch",
+    "test": "npm run build && npm run test:unit && npm run test:coverage",
+    "test:unit": "mocha --recursive --require=./test/setup.js -R spec test/index",
+    "test:coverage": "mocha --recursive --require=./test/setup.js -R mocha-text-cov test/index",
+    "demo": "superstatic demo",
+    "clean": "rimraf lib dist",
+    "prepublishOnly": "npm run build"
   },
   "keywords": [
     "react",
@@ -45,36 +40,43 @@
     "LICENSE"
   ],
   "dependencies": {
-    "@types/quill": "1.3.10",
-    "@types/react": "*",
-    "create-react-class": "^15.6.0",
+    "@types/lodash": "^4.14.146",
+    "@types/quill": "^1.3.10",
+    "@types/react": "^16.9.11",
     "lodash": "^4.17.4",
-    "prop-types": "^15.5.10",
-    "quill": "^1.3.7",
-    "react-dom-factories": "^1.0.0"
+    "quill": "^1.3.7"
   },
   "peerDependencies": {
-    "react": "^0.14.9 || ^15.3.0 || ^16.0.0"
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   },
   "devDependencies": {
+    "@types/react-dom": "^16.9.4",
+    "@types/chai": "^4.2.11",
+    "@types/mocha": "^7.0.2",
+    "@types/sinon": "^7.5.2",
     "blanket": "^1.2.3",
-    "chai": "^4.1.0",
-    "chai-enzyme": "^0.8.0",
-    "cheerio": "^1.0.0-rc.1",
-    "enzyme": "^2.9.1",
+    "chai": "^4.2.0",
+    "chai-enzyme": "^1.0.0-beta.1",
+    "cheerio": "^1.0.0-rc.3",
+    "cpx": "^1.5.0",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.15.1",
     "jsdom": "^11.0.0",
     "jsdom-global": "^3.0.2",
-    "jshint": "^2.9.4",
-    "mocha": "^3.4.2",
+    "mocha": "^6.2.2",
     "mocha-text-cov": "^0.1.1",
-    "react": "^15.6.1",
-    "react-addons-test-utils": "^15.6.0",
-    "react-dom": "^15.6.1",
-    "react-test-renderer": "^15.6.1",
-    "should": "^4.3.0",
-    "sinon": "^2.3.5",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
+    "react-test-renderer": "^16.11.0",
+    "rimraf": "^3.0.0",
+    "should": "^13.2.3",
+    "sinon": "^7.5.0",
+    "superstatic": "^6.0.4",
     "travis-cov": "^0.2.5",
-    "uglify-js": "^3.0.18",
-    "webpack": "^1.4.13"
+    "ts-loader": "^6.2.1",
+    "typescript": "^3.7.2",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10"
   }
 }
